# optimize /runOpenApi to allow pre-dereferencing OpenAPI Spec